Update typings
The prop `fill` is marked as required in the propTypes for `CircularProgress` but is optional in the Typescript type declaration. This results in Typescript not complaining when leaving the prop empty, while propTypes generates the following warning during runtime: `Failed prop type: The prop 'fill' is marked as required in 'AnimatedCircularProgress', but its value is 'undefined'.`. The fix is to change the Typescript declaration for `fill` from optional to required.